Port to cygwin.
 * m4/vl_lib_readline.m4
   Copy from ac-archive-2007.0205.
 * src/Makefile.am
   Fix gcc include search order.
 * src/dump.c
   Fix more "array subscript has type `char'" warning.
 * src/elab.c
   Fix more "array subscript has type `char'" warning.
 * src/lexer.l
   Fix more "array subscript has type `char'" warning.
   Do not increment str in toupper() which may be macro.
 * src/lib.c
   Fix more "array subscript has type `char'" warning.
 * src/nvc.c
   Fix more "array subscript has type `char'" warning.
 * src/rt/Makefile.am
   Fix gcc include search order.
 * src/sem.c
   Fix more "array subscript has type `char'" warning.
 * src/util.c
   Cygwin doesn't support user-level context switching.
   Use reference preprocessor tricks described in m4/ax_with_curses.m4.
 * test/Makefile.am
   Fix gcc include search order.

+++ src/dump.c
@@ -62,7 +62,7 @@ static void dump_expr(tree_t t)
    case T_FCALL:
       {
          const char *name = istr(tree_ident(tree_ref(t)));
-         if (isalpha(name[0]))
+         if (isalpha((uint8_t)name[0]))
             printf("%s", name);
          else
             printf("\"%s\"", name);
