+++ src/nvc.c
@@ -50,7 +50,7 @@ static ident_t to_unit_name(const char *str)
 {
    char *name = strdup(str);
    for (char *p = name; *p; p++)
-      *p = toupper(*p);
+      *p = toupper((uint8_t)*p);
 
    ident_t i = ident_prefix(lib_name(lib_work()),
                             ident_new(name), '.');
